src/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-js-asset==3.1.2
django-taggit==6.1.0
django-tinymce==4.1.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
